+++ micrometer-test/src/test/java/io/micrometer/core/instrument/ApacheAsyncHttpClient5TimingInstrumentationVerificationTests.java
@@ -17,11 +17,12 @@ package io.micrometer.core.instrument;
 
 import io.micrometer.common.lang.Nullable;
 import io.micrometer.core.instrument.binder.httpcomponents.DefaultUriMapper;
-import io.micrometer.core.instrument.binder.httpcomponents.hc5.MicrometerHttpClientInterceptor;
+import io.micrometer.core.instrument.binder.httpcomponents.hc5.ApacheHttpClientMetricsBinder;
 import org.apache.hc.client5.http.async.methods.SimpleHttpRequest;
 import org.apache.hc.client5.http.async.methods.SimpleHttpResponse;
 import org.apache.hc.client5.http.async.methods.SimpleRequestBuilder;
 import org.apache.hc.client5.http.impl.async.CloseableHttpAsyncClient;
+import org.apache.hc.client5.http.impl.async.HttpAsyncClientBuilder;
 import org.apache.hc.client5.http.impl.async.HttpAsyncClients;
 import org.apache.hc.core5.http.ContentType;
 
@@ -34,11 +35,12 @@ class ApacheAsyncHttpClient5TimingInstrumentationVerificationTests
 
     @Override
     protected CloseableHttpAsyncClient clientInstrumentedWithMetrics() {
-        MicrometerHttpClientInterceptor interceptor = new MicrometerHttpClientInterceptor(getRegistry(), Tags.empty(),
-                true, false);
-        CloseableHttpAsyncClient client = HttpAsyncClients.custom()
-            .addExecInterceptorFirst("custom", interceptor.getExecChainHandler())
-            .build();
+
+        HttpAsyncClientBuilder clientBuilder = HttpAsyncClients.custom();
+
+        ApacheHttpClientMetricsBinder.builder(getRegistry()).build().instrument(clientBuilder);
+
+        CloseableHttpAsyncClient client = clientBuilder.build();
         client.start();
         return client;
     }
